# Segment batch search