--- a/lib/api/src/grpc/proto/points.proto
+++ b/lib/api/src/grpc/proto/points.proto
@@ -117,6 +117,11 @@ message SearchPoints {
   optional uint64 offset = 9; // Offset of the result
 }
 
+message SearchBatchPoints {
+  string collection_name = 1; // Name of the collection
+  repeated SearchPoints search_points = 2;
+}
+
 message ScrollPoints {
   string collection_name = 1;
   Filter filter = 2; // Filter conditions - return only those points that satisfy the specified conditions
@@ -175,6 +180,15 @@ message ScoredPoint {
 
 message SearchResponse {
   repeated ScoredPoint result = 1;
+  double time = 2; // Time spent to process
+}
+
+message BatchResult {
+  repeated ScoredPoint result = 1;
+}
+
+message SearchBatchResponse {
+  repeated BatchResult result = 1;
   double time = 2; // Time spent to process
 }