--- a/lib/api/src/grpc/proto/points_internal_service.proto
+++ b/lib/api/src/grpc/proto/points_internal_service.proto
@@ -15,6 +15,7 @@ service PointsInternal {
   rpc CreateFieldIndex (CreateFieldIndexCollectionInternal) returns (PointsOperationResponse) {}
   rpc DeleteFieldIndex (DeleteFieldIndexCollectionInternal) returns (PointsOperationResponse) {}
   rpc Search (SearchPointsInternal) returns (SearchResponse) {}
+  rpc SearchBatch (SearchBatchPointsInternal) returns (SearchBatchResponse) {}
   rpc Scroll (ScrollPointsInternal) returns (ScrollResponse) {}
   rpc Count (CountPointsInternal) returns (CountResponse) {}
   rpc Recommend (RecommendPointsInternal) returns (RecommendResponse) {}
@@ -59,6 +60,12 @@ message DeleteFieldIndexCollectionInternal {
 message SearchPointsInternal {
   SearchPoints search_points = 1;
   uint32 shard_id = 2;
+}
+
+message SearchBatchPointsInternal {
+  string collection_name = 1;
+  repeated SearchPoints search_points = 2;
+  uint32 shard_id = 3;
 }
 
 message ScrollPointsInternal {