--- a/lib/collection/benches/hash_ring_bench.rs
+++ b/lib/collection/benches/hash_ring_bench.rs
@@ -1,3 +1,5 @@
+mod prof;
+
 use collection::hash_ring::HashRing;
 use criterion::{criterion_group, criterion_main, Criterion};
 use rand::Rng;