Another attempt to benchmark batching query.

Main question: at what level should we split batch into individual function calls?

There are 3 main options:

- On collection level - where we can simply have batch request equivalent to the list of `SearchRequest`
  - pros: Simplest to implement - no change even in internal API
  - cons: Require to perform multiple internal calls to remote shard

- On local shard level 
  - pros: 
    - No need to change code on segment level and segment holder level
    - Single request in remote shard
   - cons
     - Same filter calculated multiple times

- On segment level
  - pros:
     - Presumably smallest overhead on function calls
     - Can reuse point selection from filters
   - cons:
     - requires the most amount of changes and potential code duplication

This PR introduces benchmark to compare scenario 2 and 3.
It makes 100 single queries to 2000 stored vectors vs single batch query with 100 vectors. Dim = 100
No indexes used.
Results:

No filters:
```
batch-search-bench/search-0                                                                             
                        time:   [4.3911 ms 4.4788 ms 4.5933 ms]
batch-search-bench/search-batch-0                                                                             
                        time:   [3.7360 ms 3.8431 ms 3.9657 ms]
```

Filter which matches 1/5 of the points:
```
batch-search-bench/search-1                                                                            
                        time:   [13.916 ms 14.146 ms 14.397 ms]
batch-search-bench/search-batch-1                                                                             
                        time:   [1.1711 ms 1.1870 ms 1.2080 ms]
```

Filter which matches all data points (to estimate filter overhead):
```
batch-search-bench/search-2                                                                            
                        time:   [19.268 ms 19.852 ms 20.570 ms]
batch-search-bench/search-batch-2                                                                            
                        time:   [5.4323 ms 5.4829 ms 5.5404 ms]
```