@agourlay 2022-08-15
## Todos

- tackle code duplication
- test batched search result vs regular search
- benchmark HNSW index search vs master
- refactor `query_points` in `struct_payload_index` to remove some Iterators